@@ -11,21 +11,18 @@ declare(strict_types=1);
 
 namespace Bzga\BzgaBeratungsstellensuche\Domain\Serializer\NameConverter;
 
-use Bzga\BzgaBeratungsstellensuche\Events;
+use Bzga\BzgaBeratungsstellensuche\Events\ExtendMapNamesEvent;
 use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
+use TYPO3\CMS\Core\EventDispatcher\EventDispatcher;
 use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\SignalSlot\Dispatcher;
 
 /**
  * @author Sebastian Schreiber
  */
 class BaseMappingNameConverter extends CamelCaseToSnakeCaseNameConverter
 {
-    /**
-     * @var Dispatcher
-     */
-    protected $signalSlotDispatcher;
+    private readonly EventDispatcher $eventDispatcher;
 
     /**
      * Mapping of names, left side incoming names in xml|array, right side name for object
@@ -46,25 +43,19 @@ class BaseMappingNameConverter extends CamelCaseToSnakeCaseNameConverter
      *
      * @param array|null $attributes
      * @param bool $lowerCamelCase
-     * @param Dispatcher|null $signalSlotDispatcher
      */
-    public function __construct(array $attributes = null, $lowerCamelCase = true, Dispatcher $signalSlotDispatcher = null)
+    public function __construct(array $attributes = null, $lowerCamelCase = true, ?EventDispatcher $eventDispatcher = null)
     {
         parent::__construct($attributes, $lowerCamelCase);
 
-        if (!$signalSlotDispatcher instanceof Dispatcher) {
-            $signalSlotDispatcher = GeneralUtility::makeInstance(Dispatcher::class);
-        }
+        $this->eventDispatcher = $eventDispatcher??GeneralUtility::makeInstance(EventDispatcher::class);
 
-        $this->signalSlotDispatcher = $signalSlotDispatcher;
-
-        $this->emitMapNamesSignal();
+        $event = new ExtendMapNamesEvent($this->mapNames, []);
+        $event = $this->eventDispatcher->dispatch($event);
+        $this->addAdditionalMapNames($event->getExtendedMapNames());
         $this->mapNamesFlipped();
     }
 
-    /**
-     * @param array $mapNames
-     */
     public function addAdditionalMapNames(array $mapNames): void
     {
         ArrayUtility::mergeRecursiveWithOverrule($this->mapNames, $mapNames);
@@ -78,23 +69,13 @@ class BaseMappingNameConverter extends CamelCaseToSnakeCaseNameConverter
 
     /**
      * @param array|string|null $propertyName
-     * @return mixed|string|null
      */
-    public function denormalize($propertyName)
+    public function denormalize($propertyName): string
     {
         if (isset($this->mapNames[$propertyName])) {
             $propertyName = GeneralUtility::underscoredToLowerCamelCase($this->mapNames[$propertyName]);
         }
 
         return $propertyName;
-    }
-
-    protected function emitMapNamesSignal(): void
-    {
-        $signalArguments = [];
-        $signalArguments['extendedMapNames'] = [];
-
-        $mapNames = $this->signalSlotDispatcher->dispatch(static::class, Events::SIGNAL_MAP_NAMES, $signalArguments);
-        $this->addAdditionalMapNames($mapNames['extendedMapNames']);
     }
 }
