@@ -27,7 +27,7 @@ class ItemsProcFunc
         $templateLayouts = $templateLayoutsUtility->getAvailableTemplateLayouts($config['row']['pid']);
         foreach ($templateLayouts as $layout) {
             $additionalLayout  = [
-                htmlspecialchars($GLOBALS['LANG']->sL($layout[0])),
+                htmlspecialchars((string)$GLOBALS['LANG']->sL($layout[0])),
                 $layout[1],
             ];
             $config['items'][] = $additionalLayout;
@@ -41,7 +41,7 @@ class ItemsProcFunc
         $formFields = $formFieldsUtility->getAvailableFormFields();
         foreach ($formFields as $formField) {
             $additionalFormField = [
-                htmlspecialchars($GLOBALS['LANG']->sL($formField[0])),
+                htmlspecialchars((string)$GLOBALS['LANG']->sL($formField[0])),
                 $formField[1],
             ];
             $config['items'][]   = $additionalFormField;
