@@ -15,19 +15,20 @@ use Bzga\BzgaBeratungsstellensuche\Domain\Manager\AbstractManager;
 use Bzga\BzgaBeratungsstellensuche\Domain\Model\AbstractEntity;
 use Bzga\BzgaBeratungsstellensuche\Domain\Model\Category;
 use Bzga\BzgaBeratungsstellensuche\Domain\Model\Entry;
-use Bzga\BzgaBeratungsstellensuche\Events;
+use Bzga\BzgaBeratungsstellensuche\Events\AfterImportEvent;
+use Bzga\BzgaBeratungsstellensuche\Events\BeforeImportEvent;
 use SimpleXMLIterator;
 use Traversable;
 
 /**
  * @author Sebastian Schreiber
  */
-class XmlImporter extends AbstractImporter
+class XmlImporter extends AbstractImporter implements \Stringable
 {
     /**
      * @var string
      */
-    public const FORMAT = 'xml';
+    final public const FORMAT = 'xml';
 
     private ?int $pid = null;
 
@@ -41,10 +42,10 @@ class XmlImporter extends AbstractImporter
 
         $this->sxe = new SimpleXMLIterator($content);
 
-        $this->emitImportSignal(Events::PRE_IMPORT_SIGNAL);
+        $this->eventDispatcher->dispatch(new BeforeImportEvent($this, $this->sxe, $this->pid, $this->serializer));
 
         // Import beratungsarten
-        $this->convertRelations($this->sxe->beratungsarten->beratungsart, $this->categoryManager, Category::class, $pid);
+        $this->convertRelations($this->categoryManager, Category::class, $pid, $this->sxe->beratungsarten->beratungsart);
         $this->categoryManager->persist();
 
         $this->entries = $this->sxe->entrys->entry;
@@ -68,12 +69,12 @@ class XmlImporter extends AbstractImporter
 
     public function persist(): void
     {
+        $this->eventDispatcher->dispatch(new AfterImportEvent($this, $this->sxe, $this->pid, $this->serializer));
         // In the end we are calling all the managers to persist, this saves a lot of memory
-        $this->emitImportSignal(Events::POST_IMPORT_SIGNAL);
         $this->entryManager->persist();
     }
 
-    private function convertRelations(Traversable $relations = null, AbstractManager $manager, string $relationClassName, int $pid): void
+    private function convertRelations(AbstractManager $manager, string $relationClassName, int $pid, Traversable $relations = null): void
     {
         if ($relations instanceof Traversable) {
             foreach ($relations as $relationData) {
@@ -81,12 +82,6 @@ class XmlImporter extends AbstractImporter
             }
         }
     }
-
-    private function emitImportSignal(string $signal): void
-    {
-        $this->signalSlotDispatcher->dispatch(static::class, $signal, [$this, $this->sxe, $this->pid, $this->serializer]);
-    }
-
     private function convertRelation(AbstractManager $manager, string $relationClassName, int $pid, \SimpleXMLIterator $relationData): void
     {
         $externalId = (integer)$relationData->index;
