@@ -33,7 +33,7 @@ class TemplateLayout implements SingletonInterface
 
         // Add TsConfig values
         foreach ($this->getTemplateLayoutsFromTsConfig($pageUid) as $templateKey => $title) {
-            if (\str_starts_with($title, '--div--')) {
+            if (\str_starts_with((string)$title, '--div--')) {
                 [$templateKey, $title] = GeneralUtility::trimExplode(',', $title, true, 2);
             }
             $templateLayouts[] = [$title, $templateKey];
