@@ -12,10 +12,12 @@ declare(strict_types=1);
 namespace Bzga\BzgaBeratungsstellensuche\View\Entry;
 
 use Bzga\BzgaBeratungsstellensuche\Domain\Model\Entry;
-use TYPO3\CMS\Extbase\Mvc\View\AbstractView;
+use TYPO3Fluid\Fluid\View\ViewInterface;
 
-final class AutocompleteJson extends AbstractView
+final class AutocompleteJson implements ViewInterface
 {
+    protected $variables = [];
+
     public function render(): string
     {
         /** @var Entry[] $entries */
@@ -25,15 +27,39 @@ final class AutocompleteJson extends AbstractView
         $suggestions = [];
 
         foreach ($entries as $entry) {
-            if (\str_starts_with($entry->getCity(), $q)) {
+            if (\str_starts_with($entry->getCity(), (string)$q)) {
                 $suggestions[] = $entry->getCity();
             }
 
-            if (\str_starts_with($entry->getZip(), $q)) {
+            if (\str_starts_with($entry->getZip(), (string)$q)) {
                 $suggestions[] = $entry->getZip();
             }
         }
 
         return json_encode(array_unique($suggestions), JSON_THROW_ON_ERROR);
+    }
+
+    public function assign($key, $value)
+    {
+        $this->variables[$key] = $value;
+        return $this;
+    }
+
+    public function assignMultiple(array $values)
+    {
+        foreach ($values as $key => $value) {
+            $this->assign($key, $value);
+        }
+        return $this;
+    }
+
+    public function renderSection($sectionName, array $variables = [], $ignoreUnknown = false): string
+    {
+        return '';
+    }
+
+    public function renderPartial($partialName, $sectionName, array $variables, $ignoreUnknown = false): string
+    {
+        return '';
     }
 }
