@@ -11,16 +11,43 @@ defined('TYPO3') or die();
 
 $extKey = 'bzga_beratungsstellensuche';
 
-\TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
-    $extKey,
-    'Pi1',
-    'LLL:EXT:bzga_beratungsstellensuche/Resources/Private/Language/locallang_be.xlf:pi1_title'
-);
+$pluginConfig = ['pi1', 'list', 'show'];
 
-$GLOBALS['TCA']['tt_content']['types']['list']['subtypes_excludelist']['bzgaberatungsstellensuche_pi1'] = 'recursive,select_key,pages';
-$GLOBALS['TCA']['tt_content']['types']['list']['subtypes_addlist']['bzgaberatungsstellensuche_pi1'] = 'pi_flexform';
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue(
-    'bzgaberatungsstellensuche_pi1',
-    'FILE:EXT:' . $extKey . '/Configuration/FlexForms/flexform_beratungsstellensuche.xml'
-);
-unset($extKey);
+foreach ($pluginConfig as $pluginName) {
+    $contentTypeName = 'bzgaberatungsstellensuche_' . str_replace('_', '', $pluginName);
+
+    \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
+        'BzgaBeratungsstellensuche',
+        \TYPO3\CMS\Core\Utility\GeneralUtility::underscoredToUpperCamelCase($pluginName),
+        'LLL:EXT:bzga_beratungsstellensuche/Resources/Private/Language/locallang_be.xlf:plugin.' . $pluginName . '.title',
+        'ext-bzgaberatungsstellensuche-wizard-icon',
+        'bzga_beratungsstellensuche'
+    );
+
+    \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue(
+        '*',
+        'FILE:EXT:bzga_beratungsstellensuche/Configuration/FlexForms/flexform_beratungsstellensuche.xml',
+        $contentTypeName
+    );
+
+    $GLOBALS['TCA']['tt_content']['types'][$contentTypeName]['showitem'] = '
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:general,
+            --palette--;;general,
+            --palette--;;headers,
+        --div--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:tabs.plugin,
+            pi_flexform,
+        --div--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:tabs.appearance,
+            --palette--;;frames,
+            --palette--;;appearanceLinks,
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:language,
+            --palette--;;language,
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:access,
+            --palette--;;hidden,
+            --palette--;;access,
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:categories,
+            categories,
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:notes,
+            rowDescription,
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:extended,
+    ';
+}
